+++ server/plugin/http.go
@@ -1,10 +1,13 @@
 package plugin
 
 import (
+	"fmt"
 	"net/http"
 
 	"github.com/gorilla/mux"
+	"github.com/mattermost/mattermost-server/v5/model"
 	"github.com/mattermost/mattermost-server/v5/plugin"
+	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle"
 )
 
 const (
@@ -13,7 +16,7 @@ const (
 	routeAutocomplete        = "/autocomplete"
 	subrouteFollowedProjects = "/followedProjects"
 	routeApporveJob          = "/job/approve"
-	envOverwriteEnv          = "/env/overwrite"
+	routeEnvOverwrite        = "/env/overwrite"
 )
 
 func (p *Plugin) initializeRouter() {
@@ -24,7 +27,7 @@ func (p *Plugin) initializeRouter() {
 	autocompleteRouter := p.router.PathPrefix(routeAutocomplete).Subrouter()
 	autocompleteRouter.HandleFunc(subrouteFollowedProjects, p.autocompleteFollowedProject).Methods("GET")
 	p.router.HandleFunc(routeApporveJob, p.httpHandleApprove).Methods("POST")
-	p.router.HandleFunc(envOverwriteEnv, p.httpHandleEnvOverwrite).Methods("POST")
+	p.router.HandleFunc(routeEnvOverwrite, p.httpHandleEnvOverwrite).Methods("POST")
 }
 
 // ServeHTTP allows the plugin to implement the http.Handler interface. Requests destined for the
@@ -32,4 +35,57 @@ func (p *Plugin) initializeRouter() {
 func (p *Plugin) ServeHTTP(c *plugin.Context, w http.ResponseWriter, r *http.Request) {
 	p.API.LogDebug("Request received", "URL", r.URL)
 	p.router.ServeHTTP(w, r)
+}
+
+// overwrite given env var after confirmation if already exist
+func (p *Plugin) httpHandleEnvOverwrite(w http.ResponseWriter, r *http.Request) {
+	userID := r.Header.Get("Mattermost-User-Id")
+	circleciToken, exists := p.Store.GetTokenForUser(userID, p.getConfiguration().EncryptionKey)
+	if !exists {
+		http.NotFound(w, r)
+	}
+
+	requestData := model.PostActionIntegrationRequestFromJson(r.Body)
+	if requestData == nil {
+		p.API.LogError("Empty request data", "request", r)
+		return
+	}
+
+	originalPost, appErr := p.API.GetPost(requestData.PostId)
+	if appErr != nil {
+		p.API.LogError("Unable to get post", "postID", requestData.PostId)
+	} else if _, appErr := p.API.UpdatePost(originalPost); appErr != nil {
+		// TODO : remove the button
+		p.API.LogError("Unable to update post", "postID", originalPost.Id)
+	}
+
+	responsePost := &model.Post{
+		ChannelId: requestData.ChannelId,
+		RootId:    requestData.PostId,
+		UserId:    p.botUserID,
+	}
+
+	action := fmt.Sprintf("%s", requestData.Context["Action"])
+	projectSlug := fmt.Sprintf("%v", requestData.Context["ProjectSlug"])
+	name := fmt.Sprintf("%v", requestData.Context["EnvName"])
+	val := fmt.Sprintf("%v", requestData.Context["EnvVal"])
+
+	switch action {
+	case "deny":
+		responsePost.Message = fmt.Sprintf("Did not overwrite env variable %s for project %s", name, projectSlug)
+		p.API.SendEphemeralPost(userID, responsePost)
+
+	case "approve":
+		if err := circle.AddEnvVar(circleciToken, projectSlug, name, val); err != nil {
+			p.API.LogError("Error occurred while adding environment variable", err)
+			responsePost.Message = fmt.Sprintf(":red_circle: Cannot overwrite env var %s:%s from Mattermost.", name, val)
+		} else {
+			responsePost.Message = fmt.Sprintf(":white_check_mark: Successfully added environment variable `%s=%s` for project %s", name, val, projectSlug)
+		}
+		p.API.SendEphemeralPost(userID, responsePost)
+
+	default:
+		p.API.LogError("action %s is not valid", action)
+
+	}
 }
